+++ src/extension.ts
@@ -200,12 +200,47 @@ export function activate(context: ExtensionContext) {
     );
     context.subscriptions.push(
         commands.registerTextEditorCommand("vscode-texttoolbox.ConvertDecimalToHexadecimal", () => {
-            TextManipulation.convertHexDec(TextManipulation.hexConversionType.decToHex);
+            TextManipulation.convertSelection(TextManipulation.conversionType.decToHex);
         })
     );
     context.subscriptions.push(
         commands.registerTextEditorCommand("vscode-texttoolbox.ConvertHexadecimalToDecimal", () => {
-            TextManipulation.convertHexDec(TextManipulation.hexConversionType.hexToDec);
+            TextManipulation.convertSelection(TextManipulation.conversionType.hexToDec);
+        })
+    );
+    context.subscriptions.push(
+        commands.registerTextEditorCommand("vscode-texttoolbox.convertFromBase64", () => {
+            TextManipulation.convertSelection(TextManipulation.conversionType.fromBase64);
+        })
+    );
+    context.subscriptions.push(
+        commands.registerTextEditorCommand("vscode-texttoolbox.convertToBase64", () => {
+            TextManipulation.convertSelection(TextManipulation.conversionType.toBase64);
+        })
+    );
+    context.subscriptions.push(
+        commands.registerTextEditorCommand("vscode-texttoolbox.fromHTML", () => {
+            TextManipulation.convertSelection(TextManipulation.conversionType.fromHTML);
+        })
+    );
+    context.subscriptions.push(
+        commands.registerTextEditorCommand("vscode-texttoolbox.toHTML", () => {
+            TextManipulation.convertSelection(TextManipulation.conversionType.toHTML);
+        })
+    );
+    context.subscriptions.push(
+        commands.registerTextEditorCommand("vscode-texttoolbox.encodeUri", () => {
+            TextManipulation.convertSelection(TextManipulation.conversionType.encodeUri);
+        })
+    );
+    context.subscriptions.push(
+        commands.registerTextEditorCommand("vscode-texttoolbox.decodeUri", () => {
+            TextManipulation.convertSelection(TextManipulation.conversionType.decodeUri);
+        })
+    );
+    context.subscriptions.push(
+        commands.registerTextEditorCommand("vscode-texttoolbox.decodeJWTToken", () => {
+            TextManipulation.convertSelection(TextManipulation.conversionType.JWTDecode);
         })
     );