+++ src/modules/textManipulation.ts
@@ -1,7 +1,8 @@
-import { window } from "vscode";
+import { Selection, TextEditor, window } from "vscode";
 import { getActiveEditor, getLinesFromDocumentOrSelection, getTextFromSelection, createNewEditor } from "./helpers";
 import * as os from "os";
 import * as path from "path";
+import jwt_decode from "jwt-decode";
 
 /**
  * Trim whitespaces from the active selection(s) or from the entire document
@@ -151,40 +152,158 @@ export function convertHexadecimalToDecimal(hex: string): number | undefined {
     return parseInt(hex, 16);
 }
 
-export function convertHexDec(conversionType: hexConversionType) {
+/**
+ * Conversion type
+ *
+ * @export
+ * @enum {number}
+ */
+export enum conversionType {
+    "toBase64" = "toBase64",
+    "fromBase64" = "fromBase64",
+    "toHTML" = "toHTML",
+    "fromHTML" = "fromHTML",
+    "decToHex" = "decToHex",
+    "hexToDec" = "hexToDec",
+    "encodeUri" = "encodeUri",
+    "decodeUri" = "decodeUri",
+    "JWTDecode" = "JWTDecode",
+}
+
+/**
+ * Convert the selected text to Base64
+ *
+ * @export
+ * @async
+ * @param {conversionType} conversion Conversion Type: toBase64 or fromBase64
+ * @returns {*}
+ */
+export async function convertSelection(conversion: conversionType) {
     const editor = getActiveEditor();
     if (!editor) {
         return;
     }
 
-    const selection = editor.selection;
-    if (!selection) {
+    const selections: any = editor.selections;
+    if (!selections) {
         return;
     }
 
-    const selectionText = getTextFromSelection(editor, selection);
-    if (!selectionText) {
-        return;
+    await convertSelectionInternal(editor, selections, conversion);
+}
+
+/**
+ * Convert the selected text to Base64 ro ASCII
+ *
+ * @export
+ * @async
+ * @param {TextEditor} editor The editor to convert the selection in
+ * @param {Selection} selection The Selection to convert
+ * @param {conversionType} conversion Conversion Type: toBase64 or fromBase64
+ * @returns {(Promise<string | undefined>)}
+ */
+export async function convertSelectionInternal(editor: TextEditor, selection: Selection[], conversion: conversionType): Promise<string | undefined> {
+    if (!editor) {
+        return Promise.reject();
     }
 
-    let newText: string | number | undefined;
+    editor.edit((editBuilder) => {
+        selection.forEach((s) => {
+            let textSelection = getTextFromSelection(editor, s);
+            if (!textSelection) {
+                return;
+            }
+            let convertedText: string | number | undefined;
 
-    if (conversionType === hexConversionType.decToHex) {
-        newText = convertDecimalToHexadecimal(+selectionText);
-    }
+            switch (conversion) {
+                case conversionType.toBase64:
+                    convertedText = convertToBase64(textSelection);
+                    break;
+                case conversionType.fromBase64:
+                    convertedText = convertFromBase64(textSelection);
+                    break;
+                case conversionType.toHTML:
+                    convertedText = convertToHTML(textSelection);
+                    break;
+                case conversionType.fromHTML:
+                    convertedText = convertFromHTML(textSelection);
+                    break;
+                case conversionType.decToHex:
+                    convertedText = convertDecimalToHexadecimal(+textSelection);
+                    break;
+                case conversionType.hexToDec:
+                    convertedText = convertHexadecimalToDecimal(<string>textSelection);
+                    break;
+                case conversionType.encodeUri:
+                    convertedText = encodeUri(textSelection);
+                    break;
+                case conversionType.decodeUri:
+                    convertedText = decodeUri(textSelection);
+                    break;
+                case conversionType.JWTDecode:
+                    convertedText = decodeJWTToken(textSelection);
+                    break;
 
-    if (conversionType === hexConversionType.hexToDec) {
-        newText = convertHexadecimalToDecimal(selectionText);
-    }
+                default:
+                    break;
+            }
 
-    if (newText) {
-        editor.edit((editBuilder) => {
-            editBuilder.replace(selection, newText!.toString());
+            if (convertedText) {
+                editBuilder.replace(s, convertedText!.toString());
+            }
         });
-    }
+    });
 }
 
-export enum hexConversionType {
-    "decToHex" = "decToHex",
-    "hexToDec" = "hexToDec"
+export function convertToBase64(text: string): string {
+    return Buffer.from(text, "binary").toString("base64");
 }
+export function convertFromBase64(text: string): string {
+    return Buffer.from(text, "base64").toString("binary");
+}
+
+export function convertToHTML(text: string): string {
+    return text.replace(/&/g, "&amp;").replace(/"/g, "&quot;").replace(/'/g, "&#39;").replace(/</g, "&lt;").replace(/>/g, "&gt;");
+}
+
+export function convertFromHTML(text: string): string {
+    return text
+        .replace(/&quot;/g, '"')
+        .replace(/&#39;/g, "'")
+        .replace(/&lt;/g, "<")
+        .replace(/&gt;/g, ">")
+        .replace(/&amp;/g, "&");
+}
+
+export function encodeUri(text: string): string {
+    return encodeURIComponent(text);
+}
+
+export function decodeUri(text: string): string {
+    return decodeURIComponent(text);
+}
+
+export function decodeJWTToken(token: string): string {
+    let decodedToken: jwtToken = {
+        token: token,
+        header: {
+            ...jwt_decode(token, { header: true }),
+        },
+        payload: {
+            ...jwt_decode(token),
+        },
+    };
+
+    return JSON.stringify(decodedToken, null, 4);
+}
+
+type jwtToken = {
+    token: string;
+    header: {
+        alg: string;
+        typ: string;
+    };
+    payload: {
+        [key: string]: string;
+    };
+};